requirements.txt
alabaster==0.7.12
amqp==2.5.2
amqplib==1.0.2
anyjson==0.3.3
anytree==2.8.0
appdirs==1.4.4
asn1crypto==1.3.0
astroid==2.4.1
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.1.0
bauh==0.9.4
Beaker==1.11.0
billiard==3.6.3.0
breezy==3.0.2
CacheControl==0.12.6
celery==4.4.2
celery-with-redis==3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.0
distro==1.5.0
Django==2.2.12
django-environ==0.4.5
django-oauth-toolkit==1.3.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.16
dulwich==0.20.2
factory-boy==2.12.0
Faker==0.9.1
fastimport==0.9.8
feedparser==5.2.1
filelock==3.0.12
flake8==3.8.3
flake8-django==1.1.1
GeoIP==1.3.2
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
incremental==17.5.0
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
Jade-Application-Kit==3.5.3
jedi==0.17.0
Jinja2==2.11.2
keyutils==0.6
kombu==4.6.8
lazy-object-proxy==1.4.3
lxml==4.5.1
Mako==1.1.3
mallard-ducktype==1.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
meson==0.54.2
mixer==6.1.3
more-itertools==8.3.0
msgpack==1.0.0
npyscreen==4.10.5
oauthlib==3.1.0
ordered-set==3.1.1
packaging==20.4
parso==0.7.0
patiencediff==0.1.0
pbr==5.4.5
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
ply==3.11
progress==1.5
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.19.1
pycodestyle==2.6.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.5
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
PyHamcrest==1.9.0
pyinotify==0.9.6
pylint==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
PyQtWebEngine==5.14.0
PySide2==5.15.0
pysmbc==1.0.20
pytest==5.4.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.33.1
python-pam==1.8.4
pytoml==0.1.21
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
regex==2020.5.14
rencode==1.0.6
reportlab==3.5.42
requests==2.23.0
retrying==1.3.3
rope==0.17.0
rq==1.4.2
service-identity==18.1.0
setproctitle==1.1.10
shiboken2==5.15.0
simplejson==3.17.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
stevedore==2.0.0
text-unidecode==1.2
tinycss==0.4
tinycss2==1.0.2
toml==0.10.1
traitlets==4.3.3
Twisted==20.3.0
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.23
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.12.1
yapf==0.30.0
zipp==3.1.0
zope.interface==5.1.0
